fix: refactor and card stack
